resolve API disconnects
This change resolves the plugin disconnecting from the Ecobee servers.  Even if there is an authentication error—for example, the Ecobee API servers are down—the tokens shouldn't be removed so the plugin can keep trying to refresh.  The access token should only be cleared once the refresh_token is missing.

The current code will erase the tokens any time the API servers go down, preventing any future refreshes.